% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amigoPlot.R
\name{amigoPlot}
\alias{amigoPlot}
\title{Plot Highest Correlated Variables}
\usage{
amigoPlot(df, n_top = 4, line_color = "dodgerblue")
}
\arguments{
\item{df}{A clean dataframe.}

\item{n_top}{The number of variable comparisons to plot.}

\item{line_color}{The regression line color.}
}
\value{
\strong{auto_plots} A dataframe containing the correlation, slope, and R^2 values for the highest correlated  variables.
}
\description{
\strong{amigoPlot} plots the n most highly correlated variables.
}
\examples{
# Clean the dataframe, fires
cleaned <- autoClean(fires, factor_tol = 10, drop_user_level = 0,
                  impute_user_level = 0, impute_factors = TRUE)

# Plot the highest correlated variables
plot_results <- amigoPlot(cleaned$clean_df)

plot_results
}
