% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDash.R
\name{autoDash}
\alias{autoDash}
\title{autoDash}
\usage{
autoDash(df)
}
\arguments{
\item{df}{A dataframe, mixed data types allowed.}
}
\value{
Outputs an interactive HTML dashbaoard.
}
\description{
Automatically create an interactive dashboard
}
\examples{
autoDash(finance)

}
