% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_finance}
\alias{clean_finance}
\title{clean_finance}
\format{
## `clean_finance`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{id, member_id}{loan and member identifiers}
  \item{loan_amnt}{listed amount of the loan applied for by the borrower}
  \item{funded_amnt}{total amount committed to that loan at that point in time}
  \item{funded_amnt_inv}{total amount committed by investors for that loan at that point in time}
  ...
}
}
\source{
<https://data-nifc.opendata.arcgis.com/datasets/nifc::wildland-fire-incident-locations/about.>
}
\usage{
clean_finance
}
\description{
Cleaned finance, to be deleted.
}
\keyword{datasets}
