% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTypes.R
\name{detectTypes}
\alias{detectTypes}
\title{detectTypes}
\usage{
detectTypes(df, factor_tol = NULL, user_tol = 80)
}
\arguments{
\item{df}{A dataframe with any combination of variable classes, dates must be properly classified as dates.}

\item{factor_tol}{A numeric value indicating the tolerance to automatically coerce a column to a factor. Report in percent from 0 to 100.}

\item{user_tol}{A numeric value indicating the tolerance to receive user input prompts to assist in selecting which integer columns should be coerced to categorical, date, or retain as an integer. Report in percent from 0 to 100.}
}
\value{
The dataframe with columns coerced to the appropriate classes.
}
\description{
'detectTypes' takes a raw dataframe and coerces variables to the appropriate class.
}
\examples{
# Define a random data frame with dates and categorical integers that are not properly classified
df <- fires

# View classes of the variables in the dataframe
str(df)

# Clean the dataframe using a tolerance of 50\%
clean <- detectTypes(df, factor_tol = 50)
str(clean)

}
