% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDAmigo.R
\name{EDAmigo}
\alias{EDAmigo}
\title{EDAmigo}
\usage{
EDAmigo(
  df,
  vals = "[^0-9A-Za-z.,[:space:]-]",
  special_user_level = 0,
  factor_tol = NULL,
  type_user_tol = 20,
  no_drop = FALSE,
  no_impute = FALSE,
  drop_col_tol = 50,
  drop_row_tol = NULL,
  drop_user_level = 0,
  impute_user_level = 0
)
}
\arguments{
\item{df}{A dataframe with any combination of variable classes, dates must be properly classified as dates.}

\item{vals}{A gsub formatted list of characters to keep or remove.}

\item{special_user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the removal of special characters, 1 indicates user interaction.}

\item{factor_tol}{A numeric value indicating the tolerance to automatically coerce a column to a factor. Report in percent of unique column values from 0 to 100.}

\item{type_user_tol}{A numeric value indicating the tolerance to receive user input prompts to assist in selecting which columns should be coerced to another class. Report in percent of unique column values from 0 to 100.}

\item{no_drop}{A boolean indicating whether or not to drop values, TRUE skips all dropping and proceeds to imputation.}

\item{no_impute}{A boolean indicating whether or not to impute values. TRUE skips imputation.}

\item{drop_col_tol}{A percent tolerance to automatically drop columns with percent missing values greater than or equal to this value. Report in percent from 0 to 100.}

\item{drop_row_tol}{A percent tolerance to automatically drop rows with percent missing values greater than or equal to this value.If values are provided for both columns and rows, columns will be dropped first.Report in percent from 0 to 100.}

\item{drop_user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the drop process, 1 indicates user interaction.}

\item{impute_user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the imputation process, 1 indicates user interaction.}

\item{impute_method}{A string indicating the method of imputation. If no_impute is set to TRUE, this is ignored.}

\item{impute_factors}{A boolean indicating whether or not to impute factor columns. If set to TRUE, the value occurring most frequently is applied to missing values.}
}
\description{
EDAmigo
}
