% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finance}
\alias{finance}
\title{finance}
\format{
## `finance`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{id, member_id}{loan and member identifiers}
  \item{loan_amnt}{listed amount of the loan applied for by the borrower}
  \item{funded_amnt}{total amount committed to that loan at that point in time}
  \item{funded_amnt_inv}{total amount committed by investors for that loan at that point in time}
  ...
}
}
\source{
<https://www.lendingclub.com/company/media-center.>

\keyword{datasets}
}
\usage{
finance
}
\description{
A subset of Lending Club data which documents customer loan information
}
\keyword{datasets}
