% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fires}
\alias{fires}
\title{fires}
\format{
## `fires`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{id, member_id}{loan and member identifiers}
  \item{loan_amnt}{listed amount of the loan applied for by the borrower}
  \item{funded_amnt}{total amount committed to that loan at that point in time}
  \item{funded_amnt_inv}{total amount committed by investors for that loan at that point in time}
  ...
}
}
\source{
<https://data-nifc.opendata.arcgis.com/datasets/nifc::wildland-fire-incident-locations/about.>
}
\usage{
fires
}
\description{
A subset of the National Interagency Fire Center Wildland Fire Incident Locations data set
}
\keyword{datasets}
