% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMax.R
\name{rowMax}
\alias{rowMax}
\title{Determine the maximum value of each row of a matrix. NAs are always ignored.}
\usage{
rowMax(X)
}
\arguments{
\item{X}{A numeric matrix}
}
\value{
A list containing 2 numeric vectors, 1 that has the maximum row value and the other that indicates which column the row max is found in.
}
\description{
Determine the maximum value of each row of a matrix. NAs are always ignored.
}
\examples{
X = matrix(rnorm(100), nrow = 10)
# Return the maximum value of each row
rowMax(X)$max_value
# Return which column contains the maximum value of each row
rowMax(X)$max_which_col
}
