% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMin.R
\name{rowMin}
\alias{rowMin}
\title{Determine the minimum value of each row of a matrix. NAs are always ignored.}
\usage{
rowMin(X)
}
\arguments{
\item{X}{A numeric matrix}
}
\value{
A list containing 2 numeric vectors, 1 that has the minimum row value and the other that indicates which column the row min is found in.
}
\description{
Determine the minimum value of each row of a matrix. NAs are always ignored.
}
\examples{
X = matrix(rnorm(100), nrow = 10)
# Return the minimum value of each row
rowMin(X)$min_value
# Return which column contains the minimum value of each row
rowMin(X)$min_which_col
}
