% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClean.R
\name{autoClean}
\alias{autoClean}
\title{autoClean}
\usage{
autoClean(
  df,
  dateForm = "\%m/\%d/\%Y",
  cat_tol = NULL,
  user_tol = 80,
  vals = "[^0-9A-Za-z.,[:space:]-]",
  drop_tol = NULL,
  user_level = 0
)
}
\arguments{
\item{user_level}{}
}
\description{
autoClean
}
