% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClean.R
\name{autoClean}
\alias{autoClean}
\title{Automatically Clean A Dataframe}
\usage{
autoClean(
  df,
  vals = "[^0-9A-Za-z.,[:space:]-]",
  special_user_level = 0,
  factor_tol = NULL,
  type_user_tol = 20,
  no_drop = FALSE,
  no_impute = FALSE,
  drop_col_tol = 50,
  drop_row_tol = NULL,
  drop_user_level = 1,
  impute_user_level = 1,
  impute_method = "median",
  impute_factors = FALSE
)
}
\arguments{
\item{df}{A dataframe with any combination of variable classes, dates must be properly classified as dates.}

\item{vals}{A gsub formatted list of characters to keep or remove.}

\item{special_user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the removal of special characters, 1 indicates user interaction.}

\item{factor_tol}{A numeric value indicating the tolerance to automatically coerce a column to a factor. Report in percent of unique column values from 0 to 100.}

\item{type_user_tol}{A numeric value indicating the tolerance to receive user input prompts to assist in selecting which columns should be coerced to another class. Report in percent of unique column values from 0 to 100.}

\item{no_drop}{A boolean indicating whether or not to drop values, TRUE skips all dropping and proceeds to imputation.}

\item{no_impute}{A boolean indicating whether or not to impute values. TRUE skips imputation.}

\item{drop_col_tol}{A percent tolerance to automatically drop columns with percent missing values greater than or equal to this value. Report in percent from 0 to 100.}

\item{drop_row_tol}{A percent tolerance to automatically drop rows with percent missing values greater than or equal to this value.If values are provided for both columns and rows, columns will be dropped first.Report in percent from 0 to 100.}

\item{drop_user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the drop process, 1 indicates user interaction.}

\item{impute_user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the imputation process, 1 indicates user interaction.}

\item{impute_method}{A string indicating the method of imputation. If no_impute is set to TRUE, this is ignored. Options are: median, mean, mode, zero, and random.}

\item{impute_factors}{A boolean indicating whether or not to impute factor columns. If set to TRUE, the value occurring most frequently is applied to missing values.}
}
\value{
A list with the following elements:
\itemize{
 \item \code{clean_df}: A dataframe with special characters removed, columns coerced to appropriate classes, and missing values handled according to user input.
 \item \code{special_found_replaced}: A dataframe providing details about the special characters found in each column and which characters were removed according to user input.
 \item \code{type_stats}: A dataframe providing details about which columns were coerced to another class. Includes the percent unique values per column.
 \item \code{missing_stats}: A dataframe providing details about which columns were dropped and what percent missing values were identified throughout several steps, according to user input. If imputation is selected, this will include the method of imputation performed on each column.
 \item \code{dropped_cols}: A list of column names that were dropped from the original dataframe.
 \item \code{dropped_rows}: A list of row indices that were dropped from the original dataframe.
}
}
\description{
\strong{autoClean} takes a dataframe of mixed classes and removes special characters, coerces columns to an appropriate class, and handles missing values. All dates/times should be properly defined as 'Date', 'POSIXct', or 'POSIXlt' class. The function defaults to minimal user interaction.
}
\examples{
# Clean the dataframe, fires
cleaned <- autoClean(fires, factor_tol = 10, drop_user_level = 0, impute_user_level = 0,
        impute_factors = TRUE)

# Examine output
cleaned$special_found_replaced
cleaned$type_stats
cleaned$missing_stats
cleaned$dropped_cols
cleaned$dropped_rows

}
