% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClean.R
\name{autoClean}
\alias{autoClean}
\title{autoClean}
\usage{
autoClean(
  df,
  factor_tol = NULL,
  user_tol = 30,
  vals = "[^0-9A-Za-z.,[:space:]-]",
  drop_tol = NULL,
  user_level = 0
)
}
\arguments{
\item{df}{A dataframe.}

\item{factor_tol}{A numeric value indicating the tolerance of unique integer values to automatically coerce a column to a categorical variable. Report in percent from 0 to 100.}

\item{user_tol}{A numeric value indicating the tolerance to receive user input prompts to assist in selecting which integer columns should be coerced to categorical, date, or retain as an integer. Report in percent from 0 to 100.}

\item{vals}{A gsub formatted list of characters to keep or remove.}

\item{drop_tol}{A percent tolerance to automatically drop columns with percent missing values greater than or equal to this value.}

\item{user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the process, 1 indicates user interaction.}
}
\value{
A dataframe with special characters removed and column types set to the appropriate class.
}
\description{
autoClean
}
\examples{
cleaned <- autoClean(fires, factor_tol = 10)
}
