% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxCox.R
\name{boxCox}
\alias{boxCox}
\title{Simultaneous Boxcox power transformations}
\usage{
boxCox(X, lambda = NULL, cols = NULL, alpha = 0.001, FILTER = TRUE)
}
\arguments{
\item{X}{A numeric matrix or a dataframe containing numeric columns}

\item{lambda}{Numeric value(s) indicating what power to use in the Box-Cox transformation}

\item{cols}{A vector indicating the column numbers or the names of the columns one wishes to evaluate. If NULL then all numeric columns will be evalutated.}

\item{alpha}{A numeric value used to determine the shift parameter when 0s and or negative values are detected in the data.}

\item{FILTER}{A bolean value the at}
}
\value{
The boxCox function returns a list of objects. The boxCox_Results data frame has the estimates for each column evaluated by the function. The lambda_1 vector contains all of the lambdas evaluated by the function.
}
\description{
\strong{boxCox} is a function that aims to ease the EDA process by computing the boxcox transformations for all of the specified columns in a matrix or dataframe at the same time.
}
\details{
This implementation of BoxCox uses the following form of the normal log likelihood when selecting the power transformation value lambda. \deqn{log(L(\hat{\mu} , \hat{\sigma})) = -(\frac{n}{2})(log(2\pi \hat{\sigma^2}) +1)+ n(\lambda -1)log(GM(y))} where GM is the geomtric mean.
Once the boxCox function has computed the likelihoods for all of the lambdas provided.
Often times a simple transformation is desirable for box cox so that the results can be more easily interpreted. By default, the boxCox function selects a simple value within the 95% confidence interval as the seleced power transformation. For example, if maximum log likelihood occurs at lamda = 0.05 and 0 is contained withing the CI, then lambda will be set to 0 and the log transformation will be used.
}
\examples{
test_data =
data.frame(
 X_1 = rchisq(1000, df = 1)
 ,X_2 = rchisq(1000, df = 5)
)

boxCox(test_data)$boxCox_Results
plot(density(test_data$X_1), main = "Before")
plot(density(boxCox(test_data)$transformations$X_2), main = "After")
}
