% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectMissing.R
\name{detectMissing}
\alias{detectMissing}
\title{detectMissing}
\usage{
detectMissing(df, drop_tol = NULL, user_level = 1)
}
\arguments{
\item{df}{A dataframe.}

\item{user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the removal of special characters.}
}
\description{
detectMissing
}
