% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTypes.R
\name{detectTypes}
\alias{detectTypes}
\title{detectTypes}
\usage{
detectTypes(df, factor_tol = NULL, user_tol = 30)
}
\arguments{
\item{df}{A dataframe with any combination of variable classes, dates must be properly classified as dates.}

\item{factor_tol}{A numeric value indicating the tolerance to automatically coerce a column to a factor. Report in percent from 0 to 100.}

\item{user_tol}{A numeric value indicating the tolerance to receive user input prompts to assist in selecting which columns should be coerced to another class. Report in percent from 0 to 100.}
}
\value{
'df'= The dataframe with all variables coerced, 'date_times' = A dataframe with only the date/time columns, 'numbers' = A dataframe with only the numeric columns,
 'characters' = A dataframe with only char columns, 'factors' = A dataframe with only the factor columns, 'typeStats' = A dataframe listing the percentage of unique values in each column)
}
\description{
'detectTypes' takes a dataframe and coerces variables to the appropriate class. All dates/times should be properly defined as 'Date', 'POSIXct', or 'POSIXlt' class.
}
\examples{
# Define the dataframe to be properly classified.
df <- fires

# View classes of the variables in the dataframe
str(df)

# Clean the dataframe using a tolerance of 10\%
clean <- detectTypes(df, factor_tol = 10, user_tol = 10)
str(clean$df)

# View information about the type coercions
clean$type_stats
}
