% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectType.R
\name{detectTypes}
\alias{detectTypes}
\title{detectTypes}
\usage{
detectTypes(df, dateForm = "\%m/\%d/\%Y", cat_tol = NULL, user_tol = 80)
}
\arguments{
\item{df}{A dataframe with any combination of variable classes.}

\item{dateForm}{A string, or vector of strings, indicating the format of any dates contained in the dataframe.}

\item{cat_tol}{A numeric value indicating the tolerance of unique integer values to automatically coerce a column to a categorical variable. Report in percent from 0 to 100.}

\item{user_tol}{A numeric value indicating the tolerance to receive user input prompts to assist in selecting which integer columns should be coerced to categorical, date, or retain as an integer. Report in percent from 0 to 100.}
}
\value{
The dataframe with columns coerced to the appropriate classes.
}
\description{
'detectTypes' takes a raw dataframe and coerces variables to the appropriate class.
}
\examples{
# Define a random data frame with dates and categorical integers that are not properly classified
df <- data.frame(matrix(rnorm(20), nrow = 10))
df$Date <- rep('12/01/2014', 10)
df$DOY <- rep('123', 10)
df$int2cat <- rep(c(1,2,3,4,5), 2)

# View classes of the variables in the dataframe
str(df)

# Define the format of the dates in the dataframe
dateForm <- c('\%j', "\%m/\%d/\%Y")

# Clean the dataframe using a tolerance of 1\%
clean <- detectTypes(df, dateForm, cat_tol = 1)
str(clean)

}
