% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleMissing.R
\name{handleMissing}
\alias{handleMissing}
\title{handleMissing}
\usage{
handleMissing(df, drop_col_tol = NULL, drop_row_tol = NULL, user_level = 1)
}
\arguments{
\item{df}{A dataframe.}

\item{user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the process, 1 indicates user interaction.}

\item{drop_tol}{A percent tolerance to automatically drop columns with percent missing values greater than or equal to this value.}
}
\value{
The dataframe with dropped or imputated missing values.
}
\description{
handleMissing
}
\examples{
# Define the dataframe, clear of special characters
df <- handleSpecial(fires,user_level =0)$df

# Store function results
out <- handleMissing(df, drop_col_tol = 80, user_level = 0)
str(out$df)

}
