% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleMissing.R
\name{handleMissing}
\alias{handleMissing}
\title{handleMissing}
\usage{
handleMissing(
  df,
  no_drop = FALSE,
  no_impute = FALSE,
  drop_col_tol = NULL,
  drop_row_tol = NULL,
  missing_user_level = 1
)
}
\arguments{
\item{df}{A dataframe with any combination of variable classes, dates must be properly classified as dates.}

\item{no_drop}{A boolean indicating whether or not to drop values, TRUE skips all dropping and proceeds to imputation.}

\item{no_impute}{A boolean indicating whether or not to impute values. TRUE skips imputation.}

\item{drop_col_tol}{A percent tolerance to automatically drop columns with percent missing values greater than or equal to this value. Report in percent from 0 to 100.}

\item{drop_row_tol}{A percent tolerance to automatically drop rows with percent missing values greater than or equal to this value.If values are provided for both columns and rows, columns will be dropped first.Report in percent from 0 to 100.}

\item{missing_user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the process, 1 indicates user interaction.}
}
\value{
A list with the following elements:
\itemize{
 \item \code{df}:  A dataframe with missing values dropped and/or imputed, according to user input.
 \item \code{missing_stats}: A dataframe providing details about which columns were dropped and what percent missing values were identified throughout several steps, according to user input. If imputation is selected, this will include the method of imputation performed on each column.
 \item \code{dropped_cols}: A list of column names that were dropped from the original dataframe.
 \item \code{dropped_rows}: A list of row indices that were dropped from the original dataframe.
 }
}
\description{
'handleMissing' takes a dataframe and filsl or removes missing values. All dates/times should be properly defined as 'Date', 'POSIXct', or 'POSIXlt' class. Function defaults to user interaction.
}
\examples{
# Define the dataframe, clear of special characters
df <- handleSpecial(fires, special_user_level =0)$df

# Store function results
out <- handleMissing(df, drop_col_tol = 80, missing_user_level = 0)
str(out$df)

}
