% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleSpecial.R
\name{handleSpecial}
\alias{handleSpecial}
\title{handleSpecial}
\usage{
handleSpecial(df, vals = "[^0-9A-Za-z.,[:space:]-]", user_level = 1)
}
\arguments{
\item{df}{A dataframe.}

\item{vals}{A gsub formatted list of characters to keep or remove.}

\item{user_level}{An indicator of whether the user will provide input or if the user would like to fully automate the removal of special characters, 1 indicates user interaction.}
}
\value{
The dataframe without special characters, a dataframe containing the list of all found special characters, and a dataframe with a list of all removed special characters.
}
\description{
handleSpecial
}
\examples{
out <- handleSpecial(finance, user_level = 0)
out[[2]]

}
